Converted EcosystemDatabase to TypeScript.
