+++ src/modules/ChangeTracking/db.spec.ts
@@ -1,4 +1,5 @@
 import initApi from "../../initApi";
+import { ChangeTrackingModel } from "./db";
 
 const config = {
     ChangeTracking: { enabled: true, require: ['CoreUnit'] },
@@ -7,7 +8,7 @@ const config = {
 
 it('returns a change tracking event as last activity of a core unit', async () => {
     const apiModules = await initApi(config);
-    const db = apiModules.datasource as any;
+    const db = apiModules.datasource;
 
     const cu = {
         id: 10,
@@ -15,12 +16,17 @@ it('returns a change tracking event as last activity of a core unit', async () =
         shortCode: 'EXA'
     };
 
-    const entry = await db.ChangeTracking.getCoreUnitLastActivity(cu.id, cu.code, cu.shortCode);
+    const entry = await db.module<ChangeTrackingModel>("ChangeTracking").getCoreUnitLastActivity(cu.id, cu.code, cu.shortCode);
 
-    expect(entry.event).toBe('CU_BUDGET_STATEMENT_CREATE');
-    expect(entry.params.coreUnit).toEqual({
-        id: cu.id,
-        code: cu.code,
-        shortCode: cu.shortCode
+    expect(entry?.event).toEqual('CU_BUDGET_STATEMENT_CREATE');
+    expect(entry?.params).toEqual({
+        budgetStatementId: 123,
+        month: '2022-09',
+        coreUnit: {
+            id: cu.id,
+            code: cu.code,
+            shortCode: cu.shortCode
+        }
     });
+
 });
